// mcp/src/client/generated/models/CreateItemRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * gomoking bff API
 * BFF
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface CreateItemRequestDto
 */
export interface CreateItemRequestDto {
  /**
   * アイテム名
   * @type {string}
   * @memberof CreateItemRequestDto
   */
  name: string;
  /**
   * カテゴリ
   * @type {string}
   * @memberof CreateItemRequestDto
   */
  category: CreateItemRequestDtoCategoryEnum;
  /**
   * パントリーID
   * @type {number}
   * @memberof CreateItemRequestDto
   */
  pantryId: number;
  /**
   * 数量
   * @type {number}
   * @memberof CreateItemRequestDto
   */
  quantity: number;
  /**
   * 単位
   * @type {string}
   * @memberof CreateItemRequestDto
   */
  unit: string;
  /**
   * 賞味期限
   * @type {object}
   * @memberof CreateItemRequestDto
   */
  expiresAt?: object | null;
}

/**
 * @export
 */
export const CreateItemRequestDtoCategoryEnum = {
  Food: "Food",
  Drink: "Drink",
  Snack: "Snack",
  Spice: "Spice",
  Other: "Other",
} as const;
export type CreateItemRequestDtoCategoryEnum =
  (typeof CreateItemRequestDtoCategoryEnum)[keyof typeof CreateItemRequestDtoCategoryEnum];

/**
 * Check if a given object implements the CreateItemRequestDto interface.
 */
export function instanceOfCreateItemRequestDto(value: object): value is CreateItemRequestDto {
  if (!("name" in value) || value["name"] === undefined) return false;
  if (!("category" in value) || value["category"] === undefined) return false;
  if (!("pantryId" in value) || value["pantryId"] === undefined) return false;
  if (!("quantity" in value) || value["quantity"] === undefined) return false;
  if (!("unit" in value) || value["unit"] === undefined) return false;
  return true;
}

export function CreateItemRequestDtoFromJSON(json: any): CreateItemRequestDto {
  return CreateItemRequestDtoFromJSONTyped(json, false);
}

export function CreateItemRequestDtoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CreateItemRequestDto {
  if (json == null) {
    return json;
  }
  return {
    name: json["name"],
    category: json["category"],
    pantryId: json["pantryId"],
    quantity: json["quantity"],
    unit: json["unit"],
    expiresAt: json["expiresAt"] == null ? undefined : json["expiresAt"],
  };
}

export function CreateItemRequestDtoToJSON(json: any): CreateItemRequestDto {
  return CreateItemRequestDtoToJSONTyped(json, false);
}

export function CreateItemRequestDtoToJSONTyped(
  value?: CreateItemRequestDto | null,
  ignoreDiscriminator = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    name: value["name"],
    category: value["category"],
    pantryId: value["pantryId"],
    quantity: value["quantity"],
    unit: value["unit"],
    expiresAt: value["expiresAt"],
  };
}